mvcMainPage Refactoring
