+++ .env
@@ -1,0 +1,1 @@
+OPENAI_API_KEY="Key here! ahaaa! Bro thought he getting a free key"
