+++ README.md
@@ -1,1 +1,4 @@
 # NewsWebApp
+
+Start the server by running node app.js
+Open the server in your browser by going to localhost:3000
