--- frontend/createAccount.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Account Setup</title>
-
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-
-    <script type="text/javascript" src="./javascript/loginValidation.js"></script>
-  </head>
-
-  <body>
-    <div class="container">
-        <div class="row justify-content-center">
-          <div class="col-md-6 col-lg-4 mt-3">
-            <h1>Create New Account</h1>
-            <form class="mt-3" action="" id="signup" method="POST" onsubmit="return validateAccount(this);">
-              <div class="mb-3">
-                <label for="name" class="form-label">Name</label>
-                <input type="text" class="form-control" id="name">
-                <div id="error1" class="form-text" style="color:red" hidden="true">Your name is required.</div>
-                <br>
-                <label for="email" class="form-label">Email address</label>
-                <input type="email" class="form-control" id="email" aria-describedby="emailHelp">
-                <div id="error2" class="form-text" style="color:red" hidden="true">An email address is required.</div>
-                <br>
-                <label for="password" class="form-label">Password</label>
-                <input id="password" type="password" class="form-control">
-                <div id="error3" class="form-text" style="color:red" hidden="true">A password is required.</div>
-                <br>
-                <label for="confirm class="form-label">Confirm password</label>
-                <input id="confirm" type="password" class="form-control">
-                <div id="error4" class="form-text" style="color:red" hidden="true">The passwords you entered don't match.</div>
-              </div>
-              <button type="submit" class="btn btn-primary">Create Account</button>
-            </form>
-          </div>
-        </div>
-      </div>
-  </body>
-</html>
