--- frontend/login.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Login</title>
-
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-
-    <script type="text/javascript" src="./javascript/loginValidation.js"></script>
-  </head>
-
-  <body>
-    <div class="container">
-        <div class="row justify-content-center">
-          <div class="col-md-6 col-lg-4 mt-3">
-            <h1>Login to Portal</h1>
-            <form class="mt-3" action="" id="signup" method="POST" onsubmit="return validateLogin(this);">
-                <div class="mb-3">
-                    <label for="email" class="form-label">Email address</label>
-                    <input type="email" class="form-control" id="email" aria-describedby="emailHelp">
-                    <div id="error1" class="form-text" style="color:red" hidden="true">Email is required.</div>
-                </div>
-                
-                <div class="mb-3">
-                    <label for="password" class="form-label">Password</label>
-                    <input id="password" type="password" class="form-control">
-                    <div id="error2" class="form-text" style="color:red" hidden="true">Password is required.</div>
-                </div>
-                <button type="submit" class="btn btn-primary">Log In</button>
-            </form>
-          </div>
-        </div>
-      </div>
-  </body>
-</html>
